requirements.txt
streamlit
nltk
pypdf2
python-docx
python-pptx
openpyxl
pandas
PyPDF2
